* fpga: (139 commits)
  readermac
  fix SKIPREADLINE compilation err
  less cpu usage,  thanks @mwalker33
  iclass replay params
  hf iclass replay
  textual
  hf iclass replay works.  A bit bad in not reading AA2.  Assumes a KD mac
  iclass sniff now works
  chg: trace list color in list
  send arguments to standalone mode
  simlper select
  textual
  iceclass reader attack mode to save unique files
  add: mem spiffs wipe
  dump size bugs
  iclass sniff, still too slow
  fix the config trigger
  remove some extras
  pm3other compilation error mplicit declaration of function "Csprintf"
  style and trying a long timeout
  ...

+++ armsrc/iso14443b.h
@@ -28,7 +28,7 @@
 
 void iso14443b_setup(void);
 uint8_t iso14443b_apdu(uint8_t const *message, size_t message_length, uint8_t *response);
-uint8_t iso14443b_select_card(iso14b_card_select_t *card);
+int iso14443b_select_card(iso14b_card_select_t *card);
 uint8_t iso14443b_select_card_srx(iso14b_card_select_t *card);
 
 void SimulateIso14443bTag(uint32_t pupi);
